Refactored apropos to add functionality, to simplify the code, to fix a potentially lurking bug with non-symbolic bound names, and to paginate.
The `Exported symbols only?' question now determines whether to look
through the exports of structures available in the config package or
whether to look at the bindings visible inside a package, where
before it was merged with the package specifier in an ad-hoc way.

Nil for the package specifier now means either to look through all
structures, if exported symbols only are interesting, or to look
through the interaction environment, if not.

darcs-hash:20060622232319-85590-23108028d6269938b3eb9bcaacbddf250afebca5.gz