Refactor handling of interrupts from Emacs so that they cause an interrupt event to be sent to the session scheduler, instead of just appending some cruft onto the continuation of the thread in question with INTERRUPT-THREAD to signal a breakpoint condition.
The procedure that we add to the thread's continuation with
INTERRUPT-THREAD won't actually be run until the thread itself is
run.  For blocked threads, this means that the procedure won't be run
at all until the thread is unblocked -- which may never happen, in a
deadlock situation.  (Deadlock will usually never happen in SLIME48
because of the usual command processor always waiting for console
input.)  If we send an interrupt event to the session scheduler,
though, it can act immediately and reliably.

In order for the scheduler to push a level usefully on interrupt,
though, it must have winder & unwinder procedures for the level.
There are, for this, two new Swank session data slots, initialized in
the SLIME48 instantiation as SLDB-SEND-ACTIVATION and
SLDB-SEND-RETURN; we can't just call those procedures directly,
however, because that would introduce a circular module dependency,
and I prefer to maintain the property of independence of the session
system from the debugger mechanism.

darcs-hash:20061124080600-00fcc-ef0f21fa8c38f0fed50479ba296eef92deefd0b1.gz