Update REPL for CVS SLIME's new REPL output protocol.
This is not the best way to go about it: the new protocol allows for
the output to be generated incrementally, so that we don't need to
store it in one gigantic string.  At some point, we also want to
implement nested presentations, which are easier with the new
protocol.

darcs-hash:20070512220003-00fcc-aa9e1ffdd8c029dccacab45648089a1aea6cb2a2.gz